feat: principal details service 구현
