+++ src/main/java/goorm/reinput/user/repository/UserRepository.java
@@ -1,0 +1,13 @@
+package goorm.reinput.user.repository;
+
+import goorm.reinput.user.domain.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.Optional;
+
+@Repository
+public interface UserRepository extends JpaRepository<User, Long>{
+
+    Optional<User> findByUserId(Long userId);
+}
